sanitizers, reducing dependencies, conditional compiliation of deps, fix leaks. closes #59. (#76)
* add sanitizer tests for memory leaks and race conditions

* renice threads frequently to induce interesting interleavings

* reduce dependencies as much as possible through feature flags

* add failing leak test to travis configuration

* fix memory leaks

* improve arguments to flamerun.sh script

+++ src/log/periodic_flusher.rs
@@ -10,12 +10,12 @@ pub fn flusher(
     name: String,
     iob: Arc<IoBufs>,
     shutdown: Arc<AtomicBool>,
-    flush_every_ms: usize,
+    flush_every_ms: u64,
 ) -> std::io::Result<thread::JoinHandle<()>> {
     thread::Builder::new().name(name).spawn(move || while
         !shutdown.load(SeqCst)
     {
         iob.flush();
-        thread::sleep(Duration::from_millis(flush_every_ms as u64));
+        thread::sleep(Duration::from_millis(flush_every_ms));
     })
 }
